// src/core/Generator.d.ts
export declare enum Generator {
    BARS = "bars",
    BARS_BLOCKS = "bars blocks",
    BIG_BARS = "big bars",
    CUBES = "cubes",
    DUAL_BARS = "dualbars",
    DUAL_BARS_BLOCKS = "dualbars blocks",
    FIREWOKS = "fireworks",
    FLOWER = "flower",
    FLOWER_BLOCKS = "flower blocks",
    ORBS = "orbs",
    RING = "ring",
    RINGS = "rings",
    ROUND_WAVE = "round wave",
    SHOCKWAVE = "shockwave",
    SHINE = "shine",
    SHINE_RINGS = "shine rings",
    STAR = "star",
    STATIC = "static",
    STITCHES = "stitches",
    WAVE = "wave",
    WEB = "web"
}
export default Generator;